Move preprocess_logCPM_1kHVG() to tools.normalize
