+++ openproblems/tasks/dimensionality_reduction/methods/phate.py
@@ -1,7 +1,7 @@
 from ....tools.decorators import method
+from ....tools.normalize import preprocess_logCPM_1kHVG
 from ....tools.normalize import sqrt_cpm
 from ....tools.utils import check_version
-from .preprocessing import preprocess_logCPM_1kHVG
 
 
 def _phate(adata):
