--- openproblems/tasks/dimensionality_reduction/methods/preprocessing.py
@@ -1,15 +1,0 @@
-from ....tools.normalize import log_cpm
-from anndata import AnnData
-
-import scanpy as sc
-
-
-def preprocess_logCPM_1kHVG(adata: AnnData) -> AnnData:
-
-    log_cpm(adata)
-    sc.pp.highly_variable_genes(adata, n_top_genes=1000, flavor="cell_ranger")
-    adata = adata[:, adata.var["highly_variable"]]
-    sc.tl.pca(adata, n_comps=50, svd_solver="arpack")
-    adata.obsm["X_input"] = adata.obsm["X_pca"]
-
-    return adata
